+++ project/src/sys/gpio.c
@@ -50,7 +50,7 @@ void setPinValue(GPIO_Type* whichGPIO, uint8_t pinNumber, uint8_t value)
 	}
 }
 
-uint8_t getPinValue(GPIO_Type* whichGPIO, uint8_t pinNumber)
+uint32_t getPinValue(GPIO_Type* whichGPIO, uint8_t pinNumber)
 {
 	return (whichGPIO->PDIR & (1<<pinNumber));
 }
